studio-api/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
nltk==3.8.1
pyenchant==3.2.2
whisper-openai==1.0.0
numpy
torch==1.11.0
datasets==2.11.0
transformers==4.28.1
pandas
sentencepiece